FIXED: CID generation fixes, still does not work for MS Sans Serif ADDED: Support for QuadPoints in highligth annotations based on the work of Jonathan Verner
git-svn-id: https://svn.code.sf.net/p/podofo/code/podofo/trunk@966 e6e4d1b7-ffb3-4f57-9d89-6d52b3892af5

+++ src/PdfExtGState.cpp
@@ -71,7 +71,7 @@ void PdfExtGState::Init( void )
 
 void PdfExtGState::SetFillOpacity( float opac )
 {
-    m_pObject->GetDictionary().AddKey( "ca", PdfVariant( opac ) );
+    m_pObject->GetDictionary().AddKey( "ca", PdfVariant( static_cast<double>(opac) ) );
 }
 
 void PdfExtGState::SetStrokeOpacity( float opac )
